requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.10.1
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
crossplane==0.5.8
cryptography==41.0.4
Django==4.2.5
django-environ==0.11.2
django-extensions==3.2.3
django-revproxy @ git+https://github.com/jazzband/django-revproxy@d9914b529ed8cd0298b7a39d9f9f5e23c405e21f
gevent==23.9.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pycparser==2.21
pyOpenSSL==23.2.0
requests==2.31.0
sentry-sdk==1.32.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
zope.event==5.0
zope.interface==6.0
